project_name/requirements/base.txt
#Django==1.5.1
-e git+https://github.com/django/django.git@1.5b2#egg=django
bpython==0.12
django-braces==0.2.1
django-model-utils==1.1.0
logutils==0.3.3
